code excerpts for cookbook/effects/nested-nav
